+++ src/calc.rs
@@ -1,6 +1,8 @@
 use crate::time::HMS;
 use crate::cond::{Interval, contains_lunch_break, contains_dinner_break};
 
+pub type Meal = (bool, HMS);
+
 pub fn calc_work_time(interval: Interval) -> HMS {
     if interval.start_time < HMS::new(6, 0, 0) {
         panic!("start time must be after 6:00");
@@ -10,6 +12,41 @@ pub fn calc_work_time(interval: Interval) -> HMS {
     }
 
     let mut work_time = interval.end_time - interval.start_time;
+    let (total_exceed, _, _) = calc_total_exceed(interval);
+
+    work_time = work_time - total_exceed;
+
+    if work_time > HMS::new(12, 0, 0) {
+        work_time = HMS::new(12, 0, 0);
+    }
+
+    work_time
+}
+
+pub fn calc_minimum_finish_time(start: HMS) -> HMS {
+    calc_hope_finish_time(start, HMS::new(4,0,0))
+}
+
+pub fn calc_hope_finish_time(start: HMS, hope_work_time: HMS) -> HMS {
+    let mut finish = start + hope_work_time;
+    let mut work_time = calc_work_time(Interval::new(start, finish));
+
+    loop {
+        if work_time < hope_work_time {
+            finish = finish + (hope_work_time - work_time);
+            work_time = calc_work_time(Interval::new(start, finish));
+        } else if work_time > hope_work_time {
+            finish = finish - (work_time - hope_work_time);
+            work_time = calc_work_time(Interval::new(start, finish));
+        } else {
+            break;
+        }
+    }
+
+    finish
+}
+
+pub fn calc_total_exceed(interval: Interval) -> (HMS, Meal, Meal) {
     let mut total_exceed = HMS::new(0, 0, 0);
     let lunch = contains_lunch_break(interval);
     let dinner = contains_dinner_break(interval);
@@ -17,35 +54,23 @@ pub fn calc_work_time(interval: Interval) -> HMS {
 
     match lunch {
         (true, exceed) => {
-            println!("\tLunch break: {}", exceed);
             total_exceed = total_exceed + exceed;
         }
-        _ => {
-            println!("\tNo Lunch break");
-        }
+        _ => (),
     }
 
     match dinner {
         (true, exceed) => {
-            println!("\tDinner break: {}", exceed);
             total_exceed = total_exceed + exceed;
         }
-        _ => {
-            println!("\tNo Dinner break");
-        }
+        _ => (),
     }
 
-    if total_exceed < half_hour {
-        println!("\tTotal break: {} is less than half hour", total_exceed);
-        work_time = work_time - half_hour;
+    total_exceed = if total_exceed < half_hour {
+        half_hour
     } else {
-        work_time = work_time - total_exceed;
-    }
+        total_exceed
+    };
 
-    if work_time > HMS::new(12, 0, 0) {
-        println!("\tWork time: {} can't larger than 12 hours", work_time);
-        work_time = HMS::new(12, 0, 0);
-    }
-
-    work_time
+    (total_exceed, lunch, dinner)
 }