+++ RELEASES.md
@@ -1,3 +1,13 @@
+# Ver 0.2.0
+
+* Add calculation options
+    * Work time
+    * Minimum finish time
+    * Calculate finish time
+* Add dependency - [`dialoguer`](https://github.com/mitsuhiko/dialoguer)
+
+-----
+
 # Ver 0.1.1
 
 * Accumulated time 계산